// frontend/src/generated/model/shopping-list-dto.ts
/**
 * merged spec
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ShoppingListIngredientDto } from './shopping-list-ingredient-dto';
import { ShoppingListStatus } from './shopping-list-status';


export interface ShoppingListDto { 
    id: number;
    /**
     * Name of the shopping list
     */
    name: string;
    status: ShoppingListStatus;
    /**
     * ID of the organization managing the shopping list
     */
    organizationId: number;
    /**
     * ID of the menu the shopping list was created from
     */
    menuId?: number;
    /**
     * List of ingredients included in the shopping list
     */
    ingredients: Array<ShoppingListIngredientDto>;
}