// frontend/src/generated/model/shopping-list-edit-dto.ts
/**
 * merged spec
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ShoppingListIngredientEditDto } from './shopping-list-ingredient-edit-dto';
import { ShoppingListStatus } from './shopping-list-status';


export interface ShoppingListEditDto { 
    status?: ShoppingListStatus;
    /**
     * List of ingredients that are already in the shopping list, but need their checkedBy property updated
     */
    ingredients?: Array<ShoppingListIngredientEditDto>;
}