// frontend/src/generated/model/shopping-list-ingredient-dto.ts
/**
 * merged spec
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { IngredientUseDto } from './ingredient-use-dto';
import { IngredientCategory } from './ingredient-category';


export interface ShoppingListIngredientDto { 
    id: number;
    ingredient: IngredientUseDto;
    /**
     * Indicates if an ingredient has been checked
     */
    isChecked: boolean;
    /**
     * The person who checked the ingredient
     */
    checkedBy: string;
    category: IngredientCategory;
}