// frontend/src/generated/model/shopping-list-ingredient-edit-dto.ts
/**
 * merged spec
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ShoppingListIngredientEditDto { 
    id: number;
    /**
     * If the ingredient is checked off on the shopping list
     */
    checked: boolean;
}